Makes hopper servos run out of one port on controller
Even though this is a terrible idea, I have done it anyway. The idea is that you will run a different opmode depending on your alliance color for that match, and also switch out the servo wire that goes into port 3 on the controller. This is a bad idea, but I'm not in charge, so guess what.