Activating our OpMode for use
We have to either delete the @Disabled tag, or just comment it out in
order for our Robot Controller to advertise to the Driver Station that
an OpMode is available for use.  I commented out the tag by using the
two slashes // which java uses to mark areas as a comment.