One thumb driving is better than two...
It seems a waste to tie up both of our thumb sticks to move the
left/right motors in tank mode.  How about using just the left thumb
stick and looking at both the y (up/down axis) and the x (left/right)
axis to determine which motors should be powered and in what direction.