Add a limit switch
Last year's Pushbot had a simple push-button switch installed that could
help limit the movement of the arm and protect our robot. from
over-extending itself.  How about we enable that in software to prevent
the arm from moving too high?  To do this, we first need to import the
TouchSensor class from the robot core software.  We then need to create
a TouchSensor object (called limitSwitch) that reads the value of the
sensor from the robot.  Finally, in our loop function, we should make
sure we can only raise the arm if the limitSwitch value is 0 (ie not
triggered).