Read the description.
Got rid of the servos in all of the programs since we got rid of the
claws. This keeps all of the opmodes using HardwarePushbot from crashing
when they can't find our servo controller module. Also fixed part of
Alyssa's OpMode so it turns instead of just pausing and driving forward.