BUG: VisualNavigation integrated with iRadsAutoOpMode
Builds, but has a timeout issue.
In the iterative OpMode, the init() function is only given
5 seconds to return before an error is thrown.
Unfortunately, the vuforia initializatin routine takes about
7 seconds.
I've attempted to modify this timeout in the class properties,
but it doesn't seem to have an effect.  It could be that the parent
class (in a jar) needs to be modified.
This CAN be run from android studio in debug mode, as this seems to
suspend timeout errors.
For the time being, the linear opMode may be the simplest way to
deal with the vuforia initialization time constraints.

iRadsAutoOpMode_Linear copied from template without changes.