Add Hardware_iRads Class, uninitialized
The Hardware_iRads class, which describes our robot configuration, has been instantiated as 'robot.'
However, in the initialization block, robot.init() is commented out, so the object is not initialized. This allows the OpMode to be run on the phone without being connected to the Robot for Visual Navigation testing.  

To use with robot hardware, uncomment
robot.init(hardwareMap);
in the initialization block (line 63).