VisualNavigation: auto update timestamp with setLastLocation()
Instead of requiring lastLocationUpdateTime to be updated
manually whenever lastLocation is updated, this has been automated
with a SET method.
lastLocation is now PRIVATE, and can be accessed with getLastLocation()
and setLastLocation().
The set method assigns the current time to lastLocationUpdateTime
every time it is called to keep it valid.