TeleOp for robot
This is a teleop for the competition robot.
Anything referring to the harvester or "midMotor" is commented out until the robot has these motors.

Drive the robot with gamepad1's left and right sticks.
Using team8200's (Steel Hawks) code, control the speed and direction of the harvester with the triggers and bumpers.
Turn the motor in the middle of the robot (behind the harvester) on with (dpad) up, and turn it off with (dpad) down.

This hasn't been tested