Adding in a class for enabling/disabling the I2C-based sensors. Reference the following post: http://ftcforum.usfirst.org/showthread.php?7724-Gyro-Read-Latency-Development-team-update&p=33112&viewfull=1#post33112 for details on how to use this class.  Their example assumes you'll have this stuff in your opmode class, but now that we are using the shared 'VelocityVortexBot' class, you can add the new variables there.
Specifically, you need to add a FtcI2cDeviceState variable for each of the two color sensors, the gyro, and the distance sensor in the VelocityVortexBot class.  Then, add setEnabled(true) / setEnabled(false) calls at the top and bottom of your new functions (GyroTurn, DistanceDrive, etc) for the sensors that you will be using in that method.  The goal is to leave