Updated the FtcI2cDeviceState class with patches from the originator at TitanRobotics and updated the TODO in our InvadersVelocityVortexBot class to specify that the FtcI2cDeviceStateClass doesn't support the Modern Robotics ultrasonic range sensor.
Also: Look at this forum post regarding calling the setPower() functions on your motors in a while loop.  It's good advice.  There is no need to keep telling the motor the same power setting over and over again while reading a distance, color, or gyro sensor.  Instead, just set the power to the wheels once, then query the sensors in your while loop waiting for the conditions to change.
http://ftcforum.usfirst.org/showthread.php?7724-Gyro-Read-Latency-Development-team-update&p=33184&viewfull=1#post33184