Added a method dealing with Thread.sleep
I was informed from 8200 (Steel Hawks) that the sleep method will NOT work, so won't be used in autonomous mode.
Although this is a TEST teleop mode, and I'm kind of obsessed now at making use of the sleep method. SO IT MAY NOT WORK.

This "wait" method handles the thrown InterruptedExecption error.