Added new untested op modes for new bot
These autonomous op modes' main purpose is to shoot particles (up to 2) and  park in the center vortex.
And then there's the "trigger" servo. Lets just see if it doesn't break in the end. We need to see what the range for the servo has to be for its PROPER code configuration.